Add 'foreign' backend
